// src/pages/Landing/Types.d.ts
export type ArtistType = {
  _id: string;
  name: string;
  avatarUrl: string;
  color: string;
  orbitWidth: number;
  biography: string;
};

export type SocialMediaType = {
  name: string;
  iconUrl: string;
  link: string;
};

export type BandType = {
  logoUrl: string;
  information: string;
};